+++ my-canister-dapp-js/canister-dashboard-frontend/src/components/top-up.ts
@@ -78,7 +78,7 @@ export class TopupManager {
       const blockHeight = await this.transferToCMC(transferAmount);
 
       const cmcApi = new CMCApi();
-      const canisterIdPrincipal = await canisterId();
+      const canisterIdPrincipal = canisterId();
       const canisterIdString = canisterIdPrincipal.toString();
       await cmcApi.notifyTopUp(canisterIdString, blockHeight);
 
@@ -108,7 +108,7 @@ export class TopupManager {
   private async transferToCMC(amount: bigint): Promise<bigint> {
     const ledgerApi = new LedgerApi();
 
-    const canisterIdPrincipal = await canisterId();
+    const canisterIdPrincipal = canisterId();
     const subaccount = principalToSubaccount(canisterIdPrincipal);
 
     const cmcPrincipal = Principal.fromText(CMC_CANISTER_ID);